requirements.txt
# Core dependencies for agentic AI trader
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# LlamaIndex for RAG
llama-index>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-readers-file>=0.1.0

# OpenAI
openai>=1.0.0

# Document loading
pypdf>=3.17.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: transformers for FinBERT sentiment (comment out if not needed)
# torch>=2.0.0
# transformers>=4.30.0
